+++ packages/local_auth/local_auth_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.16
+
+* Updates androidx.fragment version to 1.5.5.
+
 ## 1.0.15
 
 * Updates androidx.fragment version to 1.5.4.