+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.4.10
+
+* Adds compatibilty with version 6.0 of the platform interface.
+
 ## 2.4.9
 
 * Fixes file URI construction.