+++ packages/video_player/video_player_android/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 2.3.10
 
+* Adds compatibilty with version 6.0 of the platform interface.
 * Fixes file URI construction in example.
 * Updates code for new analysis options.
 * Updates code for `no_leading_underscores_for_local_identifiers` lint.